requirements.txt
regex==2021.8.3
lxml==4.6.3
numpy==1.21.1
python-dateutil==2.8.2
pandas==1.3.1
requests==2.26.0
bs4==0.0.1
flask==2.2.2
flask_cors==3.0.10
gunicorn==20.1.0
pymongo==4.3.3